code written to find near by posts

+++ src/main/java/com/hangout/core/post_api/dto/GetPostsDTO.java
@@ -1,5 +1,5 @@
 package com.hangout.core.post_api.dto;
 
-public record GetPostsDTO(Double lat, Double lon, Integer searchRadius, Integer pageNumber) {
+public record GetPostsDTO(Double lat, Double lon, Double searchRadius, Integer pageNumber) {
 
 }
